services/collections-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "collections-python"
version = "0.1.0"
description = "Collections service for federated GraphQL API"
authors = [
    {name = "Kelvin Tsoi", email = "tsoikelvin@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "Django==5.0.5",
    "djangorestframework==3.14.0",
    "strawberry-graphql==0.214.0",
    "strawberry-graphql-django==0.26.0",
    "strawberry-graphql-federation==0.56.0",
    "uvicorn==0.24.0",
    "asgiref==3.7.2",
    "graphql-core>=3.2.0,<3.3.0",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.2",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.uv]
editable = false
python = "3.10"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100